NFT Purchase

+++ clips_nft/src/mint_request.rs
@@ -49,7 +49,7 @@ impl BatchMintRequest {
     /// Returns `Error::InvalidConfig` if the batch size is less than 1,
     /// or `Error::BatchLimitExceeded` if request size exceeds `config.max_batch_mint_size`.
     pub fn validate_batch_size(&self, config: &Config) -> Result<(), Error> {
-        let len = self.requests.len() as u32;
+        let len = self.requests.len();
         if len < MIN_BATCH_MINT_SIZE {
             return Err(Error::InvalidConfig);
         }
@@ -65,7 +65,7 @@ impl BatchMintRequest {
     /// Returns `Error::InvalidConfig` if batch size is 0,
     /// or `Error::BatchLimitExceeded` if request size exceeds the configured batch limit.
     pub fn validate_against_env(&self, env: &soroban_sdk::Env) -> Result<(), Error> {
-        let len = self.requests.len() as u32;
+        let len = self.requests.len();
         if len < MIN_BATCH_MINT_SIZE {
             return Err(Error::InvalidConfig);
         }
